+++ dgisim/src/character/character.py
@@ -31,6 +31,7 @@ __all__ = [
     "Character",
 
     # concretes
+    "Albedo",
     "AratakiItto",
     "Bennett",
     "Collei",
@@ -484,6 +485,76 @@ class CharacterFactory:
             equipments=self._equipments,
             statuses=self._statuses,
             elemental_aura=self._aura,
+        )
+
+
+class Albedo(Character):
+    _ELEMENT = Element.GEO
+    _WEAPON_TYPE = WeaponType.SWORD
+    _TALENT_STATUS = stt.DescentOfDivinityStatus
+    _FACTIONS = frozenset((Faction.MONDSTADT,))
+
+    _NORMAL_ATTACK_COST = AbstractDices({
+        Element.GEO: 1,
+        Element.ANY: 2,
+    })
+    _ELEMENTAL_SKILL1_COST = AbstractDices({
+        Element.GEO: 3,
+    })
+    _ELEMENTAL_BURST_COST = AbstractDices({
+        Element.GEO: 3,
+    })
+
+    @override
+    def _normal_attack(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return normal_attack_template(
+            game_state=game_state,
+            source=source,
+            element=Element.PHYSICAL,
+            damage=2,
+        )
+
+    @override
+    def _elemental_skill1(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return (
+            eft.AddSummonEffect(
+                target_pid=source.pid,
+                summon=sm.SolarIsotomaSummon,
+            ),
+        )
+
+    @override
+    def _elemental_burst(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        final_dmg = 4
+        if sm.SolarIsotomaSummon in game_state.get_player(source.pid).get_summons():
+            final_dmg += 2
+        return (
+            eft.EnergyDrainEffect(
+                target=source,
+                drain=self.get_max_energy(),
+            ),
+            eft.ReferredDamageEffect(
+                source=source,
+                target=DynamicCharacterTarget.OPPO_ACTIVE,
+                element=Element.GEO,
+                damage=final_dmg,
+                damage_type=DamageType(elemental_burst=True)
+            ),
+        )
+
+    @classmethod
+    def from_default(cls, id: int = -1) -> Self:
+        return cls(
+            id=id,
+            alive=True,
+            hp=10,
+            max_hp=10,
+            energy=0,
+            max_energy=2,
+            hiddens=stts.Statuses(()),
+            equipments=stts.EquipmentStatuses(()),
+            statuses=stts.Statuses(()),
+            elemental_aura=ElementalAura.from_default(),
         )
 
 
@@ -2310,7 +2381,7 @@ class RhodeiaOfLoch(Character):
         return tuple(effects)
 
     @classmethod
-    def from_default(cls, id: int = -1) -> RhodeiaOfLoch:
+    def from_default(cls, id: int = -1) -> Self:
         return cls(
             id=id,
             alive=True,