+++ dgisim/src/character/characters_set.py
@@ -7,6 +7,7 @@ __all__ = [
 ]
 
 _DEFAULT_CHARACTERS: list[type[Character]] = [
+    Albedo,
     AratakiItto,
     Bennett,
     Collei,